requirements.txt
asposestorage==1.0.2
h5py==2.10.0
matplotlib==3.3.1
numpy==1.18.5
Pillow==9.2.0
pytorch3d==0.3.0
tensorboardX==2.5.1
torch==1.5.1
torchvision==0.6.0a0+35d732a
trimesh==3.7.7
